3.EmbeddedModels/1_embedding_openai_query.py
from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv

load_dotenv()

embedding = OpenAIEmbeddings(model='text-embedding-3-large', dimensions=32)

result = embedding.embed_query("Delhi is the Capital of india")
# the above line will give you 32 dimensional contextual vector of the above sentence

print(str(result))